central_server/requirements.txt
fastapi
uvicorn[standard]
jinja2
aiohttp
python-multipart
redis
websockets
pyyaml
aiofiles
mcp
slowapi
